Put launch button in extensions pane
